requirements.txt
streamlit==1.32.2       # Interfaz web
#huggingface_hub==0.19.4
Pillow>=10.0.0          # Manejo de imágenes
requests>=2.31.0
google-genai>=0.3.0

#diffusers==0.24.0       # Stable Diffusion
#accelerate==0.25.0      # Optimización CPU
#safetensors==0.4.0      # Carga segura de modelos
#torch==2.0.1            # Motor de IA (CPU)
#transformers==4.35.2    # Claude /LLM